Double32_t => float. ROOT schema evolution handles this gently.

+++ DataFormats/TrackReco/interface/fillCovariance.h
@@ -4,7 +4,7 @@
 
 namespace reco {
   typedef math::Error<5>::type PerigeeCovarianceMatrix;
-  PerigeeCovarianceMatrix & fillCovariance( PerigeeCovarianceMatrix & v, const double * data );
+  PerigeeCovarianceMatrix & fillCovariance( PerigeeCovarianceMatrix & v, const float * data );
 }
   
 #endif
